Plots can be resized without a boundary
